% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateBeta.R
\name{GenerateBeta}
\alias{GenerateBeta}
\title{GenerateBeta__question3.7(Statistical Computing with R)}
\usage{
GenerateBeta(n, a, b)
}
\arguments{
\item{n}{Generate a random sample of size}

\item{a}{The first parameter of distribution function}

\item{b}{The second parameter of distribution function}
}
\value{
y The random sample has been Generated
}
\description{
Write a function to generate a random sample of size n from the Beta(a, b)
distribution by the acceptance-rejection method. Generate a random sampleof size 1000 from
the Beta(3,2) distribution. Graph the histogram of the sample with the theoretical Beta(3,2)
density superimposed
}
\examples{
y<-GenerateBeta(1000,3,2)
hist(y,prob = TRUE,ylim=c(0,2))
x <- seq(0, 1, .01)
lines(x, x^2*(1-x)^1/beta(3,2))
}
