% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateBeta.R
\name{GenerateBeta}
\alias{GenerateBeta}
\title{GenerateBeta__question3.7(Statistical Computing with R)}
\usage{
GenerateBeta(n, a, b)
}
\arguments{
\item{n}{The sise of random sample we want to generateBeta}

\item{a}{The first parameter of Beta distribution function}

\item{b}{The second parameter of Beta distribution function}
}
\value{
y The random sample has been Generated, which is a vector
}
\description{
question3.7(Statistical Computing with R):Write a function to generate a random sample of size n from the Beta(a, b)
distribution by the acceptance-rejection method. Generate a random sample of size 1000 from
the Beta(3,2) distribution. Graph the histogram of the sample with the theoretical Beta(3,2)
density superimposed

Here we use Acceptance-Rejection Method to generate a random sample, and And draw the density function of beta distribution
and compare them.
}
\examples{
y<-GenerateBeta(1000,3,2)
hist(y,prob = TRUE,ylim=c(0,2))
x <- seq(0, 1, 0.01)
lines(x, x^2*(1-x)/beta(3,2))
}
