% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEST.R
\name{GEST}
\alias{GEST}
\title{GEST__question 6.9(Statistical Computing with R)}
\usage{
GEST(distribution = "lognormal", m = 1000, n = 1000)
}
\arguments{
\item{distribution}{should be 'lognormal','uniform' or 'bernoulli',}

\item{m}{the length of random vector x}

\item{n}{Number of repeated calculations}
}
\value{
g.est the static of g.est
}
\description{
The Gini ratio is applied in economics to measure inequality in income distribution.
Note that G can be written in terms of the order statistics x
}
\examples{
GEST(distribution='lognormal',m=1e3,n=1e3)
hist(g.hat,freq = F, main = "standard lognormal" )
}
